# Conflicts:
#	ressources/js/utils/utils.js

+++ ressources/js/view/user.js
@@ -12,8 +12,9 @@ define([
     'model/user',
     'text!template/user_template.html',
     'collection/watchlist',
-    'utils/utils'
-], function(Backbone,_,$,Bootstrap, User, UserTemplate, WatchListCollection, Utils) {
+    'utils/utils',
+    'jqueryCookie'
+], function(Backbone,_,$,Bootstrap, User, UserTemplate, WatchListCollection, Utils, Cookie) {
 
     var UserView = Backbone.View.extend({
         template: _.template(UserTemplate),
@@ -26,10 +27,16 @@ define([
         },
 
         render: function (options) {
+            getToken(mycallback);
+            //console.log(result);
+
+
 
             var table = new Array();
             var that = this;
             this.user = new User({id: options.id});
+
+
 
             var watchlistCollection = new WatchListCollection();
             watchlistCollection.url = URL + '/watchlists';
@@ -39,8 +46,8 @@ define([
                         success: function (userToPrint) {
                             table.push(watchlists.toJSON());
                             table.push(userToPrint.toJSON());
+                            table.push(getCurrentUser());
                             that.$el.html(that.template({results: table}));
-
                         }
                     });
 
@@ -49,34 +56,47 @@ define([
         },
 
         followUser: function(){
-            var idData = JSON.stringify({id: $(event.target).data('id')});
-            $.ajax({
-                url: URL + '/follow',
-                type: 'POST',
-                data: idData,
-                dataType: "json",
-                contentType: 'application/json'
-            })
-                .done(function(){
-                    $('#followSuccess').fadeIn().delay(5000).fadeOut();
+            var that = this;
+            var id = $(event.target).data('id');
+            if(id != -1) {
+                var idData = JSON.stringify({id: id});
+                $.ajax({
+                    url: URL + '/follow',
+                    type: 'POST',
+                    data: idData,
+                    dataType: "json",
+                    contentType: 'application/json'
                 })
-                .fail(function(){
-                    $('#followSuccess').fadeIn().delay(5000).fadeOut();
-                })
+                    .done(function () {
+                        $('#followSuccess').fadeIn().delay(5000).fadeOut();
+                        //that.render();
+                    })
+                    .fail(function () {
+                        $('#followError').fadeIn().delay(5000).fadeOut();
+                    })
+            }
+            else{
+                $('#followError').fadeIn().delay(5000).fadeOut();
+            }
         },
 
         unfollowUser: function(){
             var id = $(event.target).data('id');
-            $.ajax({
-                url: URL + '/follow/' + id,
-                type: 'DELETE'
-            })
-                .done(function(){
-                    $('#unfollowSuccess').fadeIn().delay(5000).fadeOut();
+            if (id != -1) {
+                $.ajax({
+                    url: URL + '/follow/' + id,
+                    type: 'DELETE'
                 })
-                .fail(function(){
-                    $('#unfollowError').fadeIn().delay(5000).fadeOut();
-                })
+                    .done(function () {
+                        $('#unfollowSuccess').fadeIn().delay(5000).fadeOut();
+                    })
+                    .fail(function () {
+                        $('#unfollowError').fadeIn().delay(5000).fadeOut();
+                    })
+            }
+            else{
+                $('#unfollowError').fadeIn().delay(5000).fadeOut();
+            }
         }
 
     });
